requirements.txt
tldextract
python-whois
langchain
langchain-community
langchain-openai
openai>=1.0.0
pydantic
duckduckgo-search
streamlit
whoisit
aiohttp>=3.10.11 # not directly required, pinned by Snyk to avoid a vulnerability
anyio>=4.4.0 # not directly required, pinned by Snyk to avoid a vulnerability
requests>=2.32.2 # not directly required, pinned by Snyk to avoid a vulnerability
urllib3>=2.2.2 # not directly required, pinned by Snyk to avoid a vulnerability
zipp>=3.19.1 # not directly required, pinned by Snyk to avoid a vulnerability